Reduce SonarCloud new-code duplication for SEK-G23 executor constructors
- Move duplicated InternalInMemoryEventStore into Sekiban.Dcb.Core
  and expose it via InternalsVisibleTo.
- Add GeneralSekibanExecutorBase to share core executor, accessor, runtime
  descriptor and serialized-commit surface between WithResult/WithoutResult.
- Add InMemoryExecutorFactory (Core) and OrleansExecutorFactory (Orleans.Core)
  to consolidate accessor/inner-executor construction.
- Keep all pre-SEK-G23 public constructor overloads delegating via :this().
- Replace duplicated ExecutorBinaryCompatibilityTests with one reflection-based
  test project covering all 9 executor constructors.
- Update InternalsVisibleTo allowlist test and _core reflection probe.
