+++ dcb/src/Sekiban.Dcb.WithoutResult/Actors/GeneralSekibanExecutor.cs
@@ -14,11 +14,10 @@ namespace Sekiban.Dcb.Actors;
 ///     Wraps CoreGeneralSekibanExecutor and unwraps ResultBox, throwing exceptions on errors
 ///     This implementation uses exceptions for all error handling
 /// </summary>
-public class GeneralSekibanExecutor : ISekibanExecutor, ISerializedSekibanDcbExecutor, IExecutorRuntimeDescriptorProvider,
+public class GeneralSekibanExecutor : GeneralSekibanExecutorBase, ISekibanExecutor, ISerializedSekibanDcbExecutor,
+    IExecutorRuntimeDescriptorProvider,
     IConditionalCommandExecutor, ISerializedConditionalSekibanDcbExecutor
 {
-    private readonly CoreGeneralSekibanExecutor _core;
-    private readonly IActorObjectAccessor _actorAccessor;
     private static readonly AnonymousCommand NoCommandInstance = new();
 
     /// <summary>
@@ -39,18 +38,9 @@ public class GeneralSekibanExecutor : ISekibanExecutor, ISerializedSekibanDcbExe
         DcbDomainTypes domainTypes,
         IEventPublisher? eventPublisher = null,
         IExecutedUserProvider? executedUserProvider = null)
+        : base(eventStore, actorAccessor, domainTypes, eventPublisher, executedUserProvider)
     {
-        _actorAccessor = actorAccessor;
-        _core = new CoreGeneralSekibanExecutor(eventStore, actorAccessor, domainTypes, eventPublisher, executedUserProvider);
     }
-
-    /// <summary>
-    ///     This executor is whatever its accessor is. It has no runtime of its own — commands go wherever the accessor
-    ///     sends them — so it asks, rather than claiming. An accessor that will not say leaves this Unknown, and the
-    ///     production guard treats Unknown as unsafe, which is the only safe way to treat it.
-    /// </summary>
-    public ExecutorRuntimeDescriptor DescribeRuntime() =>
-        SekibanDcbCapabilityResolver.DescribeExecutor(_actorAccessor);
 
     /// <summary>
     ///     Execute a command with its built-in handler
@@ -182,14 +172,6 @@ public class GeneralSekibanExecutor : ISekibanExecutor, ISerializedSekibanDcbExe
     {
         return await GuardedUnwrap.UnwrapAsync(_core.GetEventStoreHeadStatusAsync(includeTotalEventCount), new BoundaryContext("ISekibanExecutor.GetEventStoreHeadStatusAsync"));
     }
-
-    public Task<ResultBox<SerializableTagState>> GetSerializableTagStateAsync(TagStateId tagStateId) =>
-        _core.GetSerializableTagStateAsync(tagStateId);
-
-    public Task<ResultBox<SerializedCommitResult>> CommitSerializableEventsAsync(
-        SerializedCommitRequest request,
-        CancellationToken cancellationToken = default) =>
-        _core.CommitSerializableEventsAsync(request, cancellationToken);
 
     /// <summary>Opt-in: execute a self-handling command with conditional (unique-key) append options.</summary>
     public async Task<ExecutionResult> ExecuteAsync<TCommand>(
